requirements.txt
Django==4.2.7
Pillow==10.1.0
pytesseract==0.3.13
opencv-python==4.8.1.78
requests==2.31.0
mysqlclient==2.2.0
google-generativeai==0.3.0
reportlab==4.0.9
